Plusieurs ameliorations

+++ Triage2.py
@@ -2,14 +2,17 @@
 
 #a faire :
 
+#utiliser les expressions regulières
 # Synchronisation sur base de donnee sur le web? Pour chercher posters et sous-titres et nom episodes
 
 
-import os
+import os, re, shutil, stat
+
 from os import rename, path
-import stat
+
 from string import capwords
-import shutil
+
+
 
 class Item_to_process:
 
@@ -25,132 +28,131 @@ class Item_to_process:
 
         self._purgatoire = '{}\purgatoire\{}'.format(source, self._nom_fichier)
 
+        self.indesirables = ['.txt', '.nfo', '.jpg', '.sfv', '.ini']
+
 
     def classify(self):
 
         #item_name, item_extention
 
         range_1_a_10 = str(list(range(0, 10)))
-        
+
         type_detecte = None
 
-        for i, n in enumerate(self._nom_fichier):
+        if self._extension not in self.indesirables:
 
-            # On detecte si c'est une serie televisee en detectant le pattern S0102
-            try:
+            for i, n in enumerate(self._nom_fichier):
 
-                if (self._nom_fichier[i] in ['s', 'S', 'E', 'e']) and (self._nom_fichier[(i + 1)] in range_1_a_10) and (self._nom_fichier[(i - 1)] in ['.', ' ']):
+                # On detecte si c'est une serie televisee en detectant le pattern S0102
+                try:
 
-                    type_detecte = 'Serie'
+                    if (self._nom_fichier[i] in ['s', 'S', 'E', 'e']) and (self._nom_fichier[(i + 1)] in range_1_a_10) and (self._nom_fichier[(i - 1)] in ['.', ' ']):
 
-                    if self._nom_fichier[i] in ['E', 'e']:
+                        type_detecte = 'Serie'
 
-                        series_episode = ('S01' + self._nom_fichier[i:(i + 3)])
+                        if self._nom_fichier[i] in ['E', 'e']:
 
-                    else:
-
-                        series_episode = ''.join(self._nom_fichier[i:(i + 6)])
-
-                    series_episode = series_episode.upper()
-
-                    series_title = ''.join(self._nom_fichier[0:(i-1)])
-
-                    series_title_clean = ''
-
-                    # On determine la nomenclature pour le dossier qui contiendra cette serie
-                    for n in series_title:
-
-                        if n != '.':
-
-                            series_title_clean = series_title_clean + n
+                            series_episode = ('S01' + self._nom_fichier[i:(i + 3)])
 
                         else:
 
-                            series_title_clean = series_title_clean + ' '
+                            series_episode = ''.join(self._nom_fichier[i:(i + 6)])
 
-                    series_title_clean = capwords(series_title_clean)
+                        series_episode = series_episode.upper()
 
+                        series_title = ''.join(self._nom_fichier[0:(i-1)])
 
+                        series_title_clean = ''
 
-                    seasons_folder = 'Season {}'.format(series_episode[1:3])
+                        # On determine la nomenclature pour le dossier qui contiendra cette serie
+                        for n in series_title:
 
-                    #On determine le path de destination pour l'item
-                    series_pathto = ('{}\{}\{}\{} {}{}'.format(dossier_series,series_title_clean, seasons_folder, series_title_clean, series_episode, self._extension))
+                            if n != '.':
 
-                    # Si le dossier qui contient la saison n'existe pas, on le cree et finalement on deplace vers celui-ci
-                    if os.path.isdir('{}\{}\{}'.format(dossier_series, series_title_clean, seasons_folder)):
+                                series_title_clean = series_title_clean + n
 
-                        print(self._path_complet + '-->' + series_pathto)
-                        if simulation == False:
-                            rename(self._path_complet, series_pathto)
+                            else:
 
-                    else:
+                                series_title_clean = series_title_clean + ' '
 
-                        os.makedirs('{}\{}\{}'.format(dossier_series, series_title_clean, seasons_folder))
-
-                        print(self._path_complet + ' --> ' + series_pathto)
-                        if simulation == False:
-                            rename(self._path_complet, series_pathto)
+                        series_title_clean = capwords(series_title_clean)
 
 
 
-            except IndexError:
-                pass
+                        seasons_folder = 'Season {}'.format(series_episode[1:3])
+
+                        #On determine le path de destination pour l'item
+                        series_pathto = ('{}\{}\{}\{} {}{}'.format(dossier_series,series_title_clean, seasons_folder, series_title_clean, series_episode, self._extension))
+
+                        # Si le dossier qui contient la saison n'existe pas, on le cree et finalement on deplace vers celui-ci
+                        if os.path.isdir('{}\{}\{}'.format(dossier_series, series_title_clean, seasons_folder)):
+
+                            print(self._path_complet + '-->' + series_pathto)
+                            if simulation == False:
+                                rename(self._path_complet, series_pathto)
+
+                        else:
+
+                            os.makedirs('{}\{}\{}'.format(dossier_series, series_title_clean, seasons_folder))
+
+                            print(self._path_complet + ' --> ' + series_pathto)
+                            if simulation == False:
+                                rename(self._path_complet, series_pathto)
+
+
+
+                except IndexError:
+                    pass
 
             # On detecte un film en trouvant une annee
 
-            try:
-                if self._nom_fichier[i] in ['1', '2'] and self._nom_fichier[i + 1] in str(list(range(0, 10))) \
-                        and self._nom_fichier[i + 2] in range_1_a_10 and self._nom_fichier[i + 3] in range_1_a_10 \
-                        and (self._nom_fichier[i - 1]) in ['.', '(', '[', '-', ' ']:
+                try:
 
-                    type_detecte = "Film"
+                    if self._nom_fichier[i] in ['1', '2'] and self._nom_fichier[i + 1] in range_1_a_10 \
+                            and self._nom_fichier[i + 2] in range_1_a_10 and self._nom_fichier[i + 3] in range_1_a_10 \
+                            and (self._nom_fichier[i - 1]) in ['.', '(', '[', '-', ' ']:
 
-                    movie_title = ''
+                        type_detecte = "Film"
 
-                    for lettre in self._nom_fichier[0:i - 1]:
+                        movie_title = ''
 
-                        if lettre == '.':
+                        #On change les points par des espaces
+                        for lettre in self._nom_fichier[0:i - 1]:
 
-                            movie_title = movie_title + ' '
+                            if lettre == '.':
 
-                        else:
+                                movie_title = movie_title + ' '
 
-                            movie_title = movie_title + lettre
+                            else:
+
+                                movie_title = movie_title + lettre
 
 
+                        movie_year = ''.join(self._nom_fichier[i:i + 4])
 
-                    print(movie_title)
+                        # dossier destination pourn les films
+                        movies_pathto = ('{}\{} ({}){}'.format(dossier_films, movie_title, movie_year, self._extension))
 
-                    movie_year = ''.join(self._nom_fichier[i:i + 4])
+                        # Deplacement du film vers le dossier film
+                        print(self._path_complet + ' --> ' + movies_pathto)
+                        if simulation == False:
+                            rename(self._path_complet, movies_pathto)
 
-                    # dossier destination pourn les films
-                    movies_pathto = ('{}\{} ({}){}'.format(dossier_films, movie_title, movie_year, self._extension))
+                except:
 
-                    # Deplacement du film vers le dossier film
-                    print(self._path_complet + ' --> ' + movies_pathto)
-                    if simulation == False:
-                        rename(self._path_complet, movies_pathto)
+                    pass
 
-            except:
+        if type_detecte == None and self._nom_fichier != os.path.basename(__file__) and self._extension not in self.indesirables:
 
-                pass
+            print('Type non-detecte, au purgatoire: ' + self._nom_fichier)
 
-        if type_detecte == None:
-
-            print('Au purgatoire: ' + self._nom_fichier)
-            print(self._purgatoire)
 
             if simulation == False:
                 os.rename(self._path_complet, self._purgatoire)
 
     def purge(self):
 
-        indesirables = ['.txt', '.nfo', '.jpg', '.sfv']
-
-
-
-        if self._extension in indesirables and self._nom_fichier not in os.listdir('{}\purgatoire'.format(source)):
+        if self._extension in self.indesirables and self._nom_fichier not in os.listdir('{}\purgatoire'.format(source)):
 
             try:
 
@@ -169,13 +171,17 @@ class Item_to_process:
 
 simulation = False
 
-source = 'd:\\downloads\\triage'
+#source = 'd:\\downloads\\triage'
 
-dossier_films = '{}\Films'.format(source)
+source = os.getcwd()
 
-dossier_series = '{}\Series'.format(source)
+dossier_films = path.join(source, 'Films')
 
-dossiers_base = ['purgatoire','Films','Series']
+dossier_series = path.join(source, 'Series')
+
+dossier_purgatoire  = path.join(source, 'Purgatoire')
+
+dossiers_base = ['Purgatoire','Films','Series']
 
 
 
@@ -194,61 +200,24 @@ for racine, directories, fichiers in root:
 
     for fichier in fichiers:
 
-        if racine[0:(len(source)+7)] != os.path.join(source,'Series') and racine[0:(len(source)+6)] != os.path.join(source,'Films'):
+
+        if racine != dossier_series and racine != dossier_films \
+                and racine !=  dossier_purgatoire:
 
             individu = Item_to_process(fichier, racine)
+            print('racine ' + racine)
+            print('dossier_series : ' + dossier_series)
             individu.purge()
+
             individu.classify()
 
 #On fait le menage des dossiers vides
 for dossier in os.listdir(source):
 
-    if dossier not in dossiers_base:
+    if dossier not in dossiers_base and dossier != os.path.basename(__file__) :
 
         print('Supression du dossier ' + (os.path.join(source, dossier)))
 
         if simulation == False:
-            #os.removedirs(os.path.join(source, dossier))
+
             shutil.rmtree(os.path.join(source, dossier))
-
-
-
-quit()
-
-
-
-for items in dossiers :
-
-    #path de l'item itere
-    path_sous_dossier = path.join(source, items)
-
-    # mode = fichier ou dossier
-    mode = os.stat(path_sous_dossier).st_mode
-
-    # Est-ce que c'est un dossier?
-    if stat.S_ISDIR(mode) and items not in dossiers_base :
-
-        #oui alors on liste son contenu
-        fichiers = os.listdir(path_sous_dossier)
-
-        #on itere et process les items contenus
-        for elements in fichiers:
-            print(elements)
-            individu = Item_to_process(elements, path_sous_dossier)
-            individu.purge()
-
-            individu.classify()
-
-        #On elimine le dossier vide qui reste
-        os.removedirs(path_sous_dossier)
-
-    # Est-ce que c'est un fichier?
-    if stat.S_ISREG(mode) and items != os.path.basename(__file__):
-
-        #oui alors on le process
-
-        individu = Item_to_process(items,source)
-        individu.purge()
-        individu.classify()
-
-
